% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_average.R
\name{ensemble_average}
\alias{ensemble_average}
\title{Creates an Average Esemble Model}
\usage{
ensemble_average(object, type = c("mean", "median"))
}
\arguments{
\item{object}{A Modeltime Table}

\item{type}{Specify the type of average ("mean" or "median")}
}
\description{
Creates an Average Esemble Model
}
