% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime_resample_accuracy.R
\name{modeltime_resample_accuracy}
\alias{modeltime_resample_accuracy}
\title{Calculate Accuracy Metrics from Modeltime Resamples}
\usage{
modeltime_resample_accuracy(
  object,
  metric_set = default_forecast_accuracy_metric_set()
)
}
\arguments{
\item{object}{a Modeltime Table with a column '.resample_results' (the output of \code{\link[=modeltime_fit_resamples]{modeltime_fit_resamples()}})}

\item{metric_set}{A \code{yardstick::metric_set()} that is used to summarize one or more
forecast accuracy (regression) metrics.}
}
\description{
This is a wrapper for \code{yardstick} that simplifies time
series regression accuracy metric calculations from
a Modeltime Table that has been resampled and fitted using
\code{\link[=modeltime_fit_resamples]{modeltime_fit_resamples()}}.
}
\examples{
library(tidymodels)
library(modeltime)
library(modeltime.ensemble)
library(timetk)
library(tidyverse)

resamples_tscv <- training(m750_splits) \%>\%
    time_series_cv(
        assess      = "2 years",
        initial     = "5 years",
        skip        = "2 years",
        slice_limit = 1
    )

m750_models_resample <- m750_models \%>\%
    modeltime_fit_resamples(
        resamples = resamples_tscv,
        control   = control_resamples(verbose = TRUE)
    )

m750_models_resample

m750_models_resample \%>\%
    modeltime_resample_accuracy() \%>\%
    table_modeltime_accuracy(.interactive = FALSE)

}
