% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble-type_sum.R
\name{type_sum.mdl_time_ensemble}
\alias{type_sum.mdl_time_ensemble}
\title{Succinct summary of Modeltime Tables}
\usage{
\method{type_sum}{mdl_time_ensemble}(x)
}
\arguments{
\item{x}{A `mdl_time_ensemble` object to summarise.}
}
\value{
A character value.
}
\description{
`type_sum` controls how objects are shown when inside tibble
 columns.
}
