% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-resamples.R
\name{unnest_resamples}
\alias{unnest_resamples}
\title{Unnests the Results of Modeltime Fit Resamples}
\usage{
unnest_resamples(object)
}
\arguments{
\item{object}{A Modeltime Table that has a column '.resample_results'}
}
\value{
Tibble with columns for '.row_id', '.resample_id', '.model_id', '.model_desc', '.pred',
'.row', and actual value name from the data set
}
\description{
An internal function used by \code{\link[=modeltime_resample_accuracy]{modeltime_resample_accuracy()}}.
}
\details{
Data Columns:
\itemize{
\item \code{.row_id} - A unique identifier to compare observations. Used in \code{ensemble_model_spec()} to compare observations by \code{.model_id}.
\item \code{.resample_id} - A unique identifier given to the resample iteration.
\item \code{.model_id} and \code{.model_desc} - Modeltime Model ID and Description
\item \code{.pred}, \code{.row}, and actual value column (name changes to name in dataset)
}
}
