% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_ensemble_weighted.R
\name{ensemble_nested_weighted}
\alias{ensemble_nested_weighted}
\title{Nested Ensemble Weighted}
\usage{
ensemble_nested_weighted(
  object,
  loadings,
  scale_loadings = TRUE,
  loading_method = c("lowest_rmse", "sequential"),
  keep_submodels = TRUE,
  model_ids = NULL,
  control = control_nested_fit()
)
}
\arguments{
\item{object}{A nested modeltime object (inherits class \code{nested_mdl_time})}

\item{loadings}{A vector of weights corresponding to the loadings}

\item{scale_loadings}{If TRUE, divides by the sum of the loadings
to proportionally weight the submodels.}

\item{keep_submodels}{Whether or not to keep the submodels in the
nested modeltime table results}

\item{model_ids}{A vector of id's (\code{.model_id}) identifying which submodels to
use in the ensemble.}

\item{control}{Controls various aspects of the ensembling process. See \code{\link[=control_nested_fit]{control_nested_fit()}}.}
}
\description{
Creates an Ensemble Model using Weighted Averaging in the
Modeltime Nested Forecasting Workflow.
}
