% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_weighted.R
\name{ensemble_weighted}
\alias{ensemble_weighted}
\title{Creates a Weighted Ensemble Model}
\usage{
ensemble_weighted(object, loadings = "auto", resamples = NULL)
}
\arguments{
\item{object}{A Modeltime Table}

\item{loadings}{Either "auto" or a vector of weights corresponding to the loadings}

\item{resamples}{NULL.}
}
\description{
Creates a Weighted Ensemble Model
}
\details{
The input to an \code{ensemble_weighted()} model is always a Modeltime Table,
which contains the models that you will ensemble.

\strong{Weighting Methods}

The average method uses an un-weighted average using \code{loadings} of either:
\itemize{
\item \code{"auto"}: Performs weighting using Penalized Regression (using and Elastic Net via \code{glmnet})
\item \verb{<vector>}:
}
}
\examples{
library(tidymodels)
library(modeltime)
library(modeltime.ensemble)
library(tidyverse)
library(timetk)

# Make an ensemble from a Modeltime Table
ensemble_fit <- m750_models \%>\%
    ensemble_weighted(loadings = c(3, 3, 1) / 7)

ensemble_fit

# Forecast with the Ensemble
modeltime_table(
    ensemble_fit
) \%>\%
    modeltime_forecast(
        new_data    = testing(m750_splits),
        actual_data = m750
    ) \%>\%
    plot_modeltime_forecast(
        .interactive = FALSE,
        .conf_interval_show = FALSE
    )

}
