% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime_resample_accuracy.R
\name{modeltime_resample_accuracy}
\alias{modeltime_resample_accuracy}
\title{Calculate Accuracy Metrics from Modeltime Resamples}
\usage{
modeltime_resample_accuracy(
  object,
  summary_fns = mean,
  metric_set = default_forecast_accuracy_metric_set()
)
}
\arguments{
\item{object}{a Modeltime Table with a column '.resample_results' (the output of \code{\link[=modeltime_fit_resamples]{modeltime_fit_resamples()}})}

\item{summary_fns}{One or more functions that is passed to \code{dplyr::across(.fns)}.
Possible values are:
\itemize{
\item NULL, to returns the resamples untransformed.
\item A function, e.g. mean.
\item A purrr-style lambda, e.g. ~ mean(.x, na.rm = TRUE)
\item A list of functions/lambdas, e.g. list(mean = mean, sd = sd)
}}

\item{metric_set}{A \code{yardstick::metric_set()} that is used to summarize one or more
forecast accuracy (regression) metrics.}
}
\description{
This is a wrapper for \code{yardstick} that simplifies time
series regression accuracy metric calculations from
a Modeltime Table that has been resampled and fitted using
\code{\link[=modeltime_fit_resamples]{modeltime_fit_resamples()}}.
}
\details{
Returns the \emph{average} accuracy metrics for each resample prediction.
}
\examples{
library(tidymodels)
library(modeltime)
library(modeltime.ensemble)
library(timetk)
library(tidyverse)

resamples_tscv <- training(m750_splits) \%>\%
    time_series_cv(
        assess      = "2 years",
        initial     = "5 years",
        skip        = "2 years",
        slice_limit = 1
    )

m750_models_resample <- m750_models \%>\%
    modeltime_fit_resamples(
        resamples = resamples_tscv,
        control   = control_resamples(verbose = TRUE)
    )

m750_models_resample

# Average
m750_models_resample \%>\%
    modeltime_resample_accuracy() \%>\%
    table_modeltime_accuracy(.interactive = FALSE)

# Mean and Standard Deviation
m750_models_resample \%>\%
    modeltime_resample_accuracy(
        summary_fns = list(mean = mean, sd = sd)
    ) \%>\%
    table_modeltime_accuracy(.interactive = FALSE)

}
